+++ ExampleApp/PACECloudSDKExample/App/AppControl.swift
@@ -23,11 +23,11 @@ class AppControl {
     private var didAuthorize = false
 
     private init() {
-        AppKit.shared.delegate = self
+        AppKit.delegate = self
     }
 
     func requestLocalApps() {
-        AppKit.shared.requestLocalApps()
+        AppKit.requestLocalApps()
     }
 
     func requestAllCoFuStations() {
@@ -37,11 +37,11 @@ class AppControl {
     }
 
     func appViewController(appUrl: String) -> UIViewController {
-        AppKit.shared.appViewController(appUrl: appUrl)
+        AppKit.appViewController(appUrl: appUrl)
     }
 
     func handleRedirectURL(_ url: URL) {
-        AppKit.shared.handleRedirectURL(url)
+        AppKit.handleRedirectURL(url)
     }
 
     func isPoiInRange(with id: String) {